Encapsulation
There is no difference between the code in Abstraction and Encapsulation, only the fact that Class in encapsulation is not implementing any interface. Encapsulation we can see that the data members and member methods are tied up in a same entity(object through classes) and hence the private members can be accessed using getter methods.